Refactor device into several files
